LevelThree tangibleEntered() logic done
logic handling for when any tangible is entered after the first (should be a class tangible) is entered. 
Error dialog if non AttributeValue tangible is entered.
Finds index of attribute class we are trying to add and updates the attributeValues array with the new attribute value from tangible object.
Error dialog displays if attributeValue entered is not in the class created

Allows user to enter any class - even the ones the user did not edit in levelTwo - since all classes by default have a size attribute (which can be removed in levelTwo)

TODO: Konva methods for drawing/setting/updating the shape.
TODO: once Konva is done change tangibleEntered in LevelThree to always return false